requirements.txt
numpy<2.0.0
Django==5.1.3
pillow
opencv-python-headless